chore(examples): update them
