+++ docs/technical-reference/interfaces/ERC725JSONSchema.md
@@ -21,7 +21,7 @@ Detailed information available on [LSP-2-ERC725YJSONSchema](https://github.com/l
 
 #### Defined in
 
-[types/ERC725JSONSchema.ts:47](https://github.com/ERC725Alliance/erc725.js/blob/7a288e2/src/types/ERC725JSONSchema.ts#L47)
+[types/ERC725JSONSchema.ts:47](https://github.com/ERC725Alliance/erc725.js/blob/00ce808/src/types/ERC725JSONSchema.ts#L47)
 
 ___
 
@@ -31,7 +31,7 @@ ___
 
 #### Defined in
 
-[types/ERC725JSONSchema.ts:48](https://github.com/ERC725Alliance/erc725.js/blob/7a288e2/src/types/ERC725JSONSchema.ts#L48)
+[types/ERC725JSONSchema.ts:48](https://github.com/ERC725Alliance/erc725.js/blob/00ce808/src/types/ERC725JSONSchema.ts#L48)
 
 ___
 
@@ -41,7 +41,7 @@ ___
 
 #### Defined in
 
-[types/ERC725JSONSchema.ts:49](https://github.com/ERC725Alliance/erc725.js/blob/7a288e2/src/types/ERC725JSONSchema.ts#L49)
+[types/ERC725JSONSchema.ts:49](https://github.com/ERC725Alliance/erc725.js/blob/00ce808/src/types/ERC725JSONSchema.ts#L49)
 
 ___
 
@@ -51,7 +51,7 @@ ___
 
 #### Defined in
 
-[types/ERC725JSONSchema.ts:50](https://github.com/ERC725Alliance/erc725.js/blob/7a288e2/src/types/ERC725JSONSchema.ts#L50)
+[types/ERC725JSONSchema.ts:50](https://github.com/ERC725Alliance/erc725.js/blob/00ce808/src/types/ERC725JSONSchema.ts#L50)
 
 ___
 
@@ -61,7 +61,7 @@ ___
 
 #### Defined in
 
-[types/ERC725JSONSchema.ts:43](https://github.com/ERC725Alliance/erc725.js/blob/7a288e2/src/types/ERC725JSONSchema.ts#L43)
+[types/ERC725JSONSchema.ts:43](https://github.com/ERC725Alliance/erc725.js/blob/00ce808/src/types/ERC725JSONSchema.ts#L43)
 
 ___
 
@@ -71,7 +71,7 @@ ___
 
 #### Defined in
 
-[types/ERC725JSONSchema.ts:44](https://github.com/ERC725Alliance/erc725.js/blob/7a288e2/src/types/ERC725JSONSchema.ts#L44)
+[types/ERC725JSONSchema.ts:44](https://github.com/ERC725Alliance/erc725.js/blob/00ce808/src/types/ERC725JSONSchema.ts#L44)
 
 ___
 
@@ -81,7 +81,7 @@ ___
 
 #### Defined in
 
-[types/ERC725JSONSchema.ts:42](https://github.com/ERC725Alliance/erc725.js/blob/7a288e2/src/types/ERC725JSONSchema.ts#L42)
+[types/ERC725JSONSchema.ts:42](https://github.com/ERC725Alliance/erc725.js/blob/00ce808/src/types/ERC725JSONSchema.ts#L42)
 
 ___
 
@@ -91,7 +91,7 @@ ___
 
 #### Defined in
 
-[types/ERC725JSONSchema.ts:45](https://github.com/ERC725Alliance/erc725.js/blob/7a288e2/src/types/ERC725JSONSchema.ts#L45)
+[types/ERC725JSONSchema.ts:45](https://github.com/ERC725Alliance/erc725.js/blob/00ce808/src/types/ERC725JSONSchema.ts#L45)
 
 ___
 
@@ -101,4 +101,4 @@ ___
 
 #### Defined in
 
-[types/ERC725JSONSchema.ts:46](https://github.com/ERC725Alliance/erc725.js/blob/7a288e2/src/types/ERC725JSONSchema.ts#L46)
+[types/ERC725JSONSchema.ts:46](https://github.com/ERC725Alliance/erc725.js/blob/00ce808/src/types/ERC725JSONSchema.ts#L46)
