+++ examples/package.json
@@ -1,12 +1,15 @@
 {
-  "name": "erc725.js examples",
+  "name": "erc725.js-examples",
   "type": "module",
   "private": true,
   "scripts": {
     "postinstall": "cd .. && npm link && cd ./examples && npm link erc725.js"
   },
   "dependencies": {
-    "web3": "1.5.0",
-    "erc725.js": "0.5.7"
+    "erc725.js": "0.5.7",
+    "web3": "1.5.0"
+  },
+  "devDependencies": {
+    "isomorphic-fetch": "^3.0.0"
   }
 }
